Prevent insertStyle tests from trampling global.document
We were just lucky that these happened to occur after our other tests which
depend on global.document (initialized in test/.setup.js). Enzyme tests will
blow up if the global document object is deleted. This change preserves the
value up front and restores it when leaving the test group.