Fix and reorganize certain tests (#41)
* Prevent insertStyle tests from trampling global.document

We were just lucky that these happened to occur after our other tests which
depend on global.document (initialized in test/.setup.js). Enzyme tests will
blow up if the global document object is deleted. This change preserves the
value up front and restores it when leaving the test group.

* Rename and reorganize certain tests

* Rename test files to match functions tested rather than filenames
* Reorganize withStyles tests into clearer groupings